Changed int's to long long's
Due to some of the input numbers being too big for int's to store, I am changing this to be able to not lose any information.

+++ day5/day5/Inventory.h
@@ -10,10 +10,12 @@ class Inventory {
 public:
 	Inventory(string);// Constructor
 
-	list<int> getLocations();// Get answer to puzzle and maps all seeds to locations and returns a list of them
+	list<long long> getLocations();// Get answer to puzzle and maps all seeds to locations and returns a list of them
+
+
 
 private:
-	list<int> seeds;
+	list<long long> seeds;
 	Almanac theMappings;
 };
